backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
docstring_parser==0.17.0
fastapi==0.115.13
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.105.0
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-monitoring==2.27.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.27.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jwt==1.4.0
numpy==2.3.1
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
structlog==25.2.0
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
unicorn==2.1.3
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
